tenkai

+++ src/main/java/com/example/revitech/controller/ChatMessagesApiController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.revitech.dto.ChatMessageDto;
+import com.example.revitech.dto.ChatMessageDto; // ★ DTOをインポート
 import com.example.revitech.service.ChatMessageService;
 
 @RestController
@@ -20,8 +20,13 @@ public class ChatMessagesApiController {
         this.chatMessageService = chatMessageService;
     }
 
+    /**
+     * ★★★ ここを修正 ★★★
+     * このAPIが返すデータの型を List<ChatMessage> から List<ChatMessageDto> に変更します。
+     */
     @GetMapping("/{roomId}")
-    public List<ChatMessageDto> getMessages(@PathVariable Integer roomId) {
+    public List<ChatMessageDto> getMessages(@PathVariable Long roomId) {
+        // 修正されたサービスメソッドを呼び出すと、自動的に正しい型のリストが返されます
         return chatMessageService.getMessagesByRoomId(roomId);
     }
 }
